+++ app/.dockerignore
@@ -1,0 +1,3 @@
+.gitignore
+node_modules
+public/hex_features_real.geojson.zip
