+++ app/Dockerfile
@@ -1,0 +1,21 @@
+# Use a base image for Node.js
+FROM node:18-alpine
+
+# Set the working directory inside the container
+WORKDIR /usr/src/app
+
+# Copy package.json and package-lock.json
+COPY package*.json ./
+
+# Install dependencies
+RUN npm install
+
+# Copy the rest of the application files
+COPY . .
+
+# Expose the port your app runs on
+EXPOSE 5173
+
+# Define the command to run your app
+CMD ["npm", "start"]
+
