--- app/features.geojson
@@ -1,1 +1,0 @@
-../data/features.geojson
