--- app/hex_features.geojson
@@ -1,1 +1,0 @@
-../data/hex_features.geojson
