--- app/hex_features_real.geojson
@@ -1,1 +1,0 @@
-../data/hex_features_real.geojson
