--- app/hex_features_real.geojson.zip
@@ -1,1 +1,0 @@
-../data/hex_features_real.geojson.zip
