+++ app/index.html
@@ -14,7 +14,7 @@
 <div id="app"></div>
 </body>
 <script type="module">
-  import {renderToDOM} from './app.tsx';
+  import {renderToDOM} from './src/app.tsx';
   renderToDOM(document.getElementById('app'));
 </script>
 </html>
