+++ app/package.json
@@ -4,11 +4,14 @@
   "private": true,
   "license": "MIT",
   "scripts": {
-    "start": "vite --open",
+    "start": "vite --host 0.0.0.0 --port 5173",
     "start-local": "vite --config ../../vite.config.local.mjs",
     "build": "vite build"
   },
   "dependencies": {
+    "@loaders.gl/core": "^4.2.0",
+    "@loaders.gl/json": "^4.2.0",
+    "@loaders.gl/zip": "^4.2.0",
     "@types/d3-scale": "^4.0.0",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
@@ -17,10 +20,7 @@
     "maplibre-gl": "^3.0.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-map-gl": "^7.1.0",
-    "@loaders.gl/core": "^4.2.0",
-    "@loaders.gl/json": "^4.2.0",
-    "@loaders.gl/zip": "^4.2.0"
+    "react-map-gl": "^7.1.0"
   },
   "devDependencies": {
     "typescript": "^4.6.0",
