+++ app/tsconfig.json
@@ -4,5 +4,8 @@
     "jsx": "react",
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true
-  }
+  },
+  "include": [
+    "src"
+  ]
 }
